requirements.txt
python-telegram-bot==20.1
python-dotenv==1.0.0
flake8==5.0.4
flake8-docstrings==1.7.0
flake8-broken-line==0.6.0
flake8-builtins==2.1.0
flake8-commas==2.1.0
flake8-comprehensions==3.10.1
flake8-import-order==0.18.2
flake8-mutable==1.2.0
flake8-quotes==3.3.2
flake8-string-format==0.3.0
flake8-variables-names==0.0.5
flake8-functions-names==0.4.0
mypy==1.0.1